implement class and status filters for test list
