+++ report-ng/app-react/src/pages/TestListPage.tsx
@@ -1,102 +1,28 @@
 import Box from "@mui/material/Box";
-import {Grid, Stack, Switch} from "@mui/material";
-import * as React from "react";
+import {Grid, Switch} from "@mui/material";
 import StatusSelectInput from "../widgets/select-input/status-select-input";
 import FormControlLabel from "@mui/material/FormControlLabel";
 import TextField from '@mui/material/TextField';
 import SearchIcon from '@mui/icons-material/Search';
 import InputAdornment from '@mui/material/InputAdornment';
-import {useSearchParams} from "react-router-dom";
-import ReportChip from "../widgets/report-chip/report-chip";
 import MultiSelectInput from "../widgets/select-input/multi-select-input";
-import {StatusService} from "../model/status-service";
-import type { ResultStatus } from "../model/status-service";
+import {useTestListFilters} from "../hooks/useTestListFilters";
+import SelectedFiltersChips from "../components/selected-filter-chips";
 
 const TestListPage = () => {
-    const statusMenuItems = StatusService.getRelevantStatuses();
 
-    const classMenuItems: { value: string, label: string }[] = [{
-        value: "SimpleTest2",
-        label: "SimpleTest2"
-    }, {value: "SimpleTest", label: "SimpleTest"}]
-
-    const [configurationMethodsChecked, setConfigurationMethodsChecked] = React.useState(false);
-    const [searchParams, setSearchParams] = useSearchParams();
-
-    const statusParam = searchParams.get("status");
-    const classParam = searchParams.get("class");
-
-    const selectedStatuses: ResultStatus[] = statusParam
-        ? statusParam
-            .split("~")
-            .map(statusKey => StatusService.getStatusByKey(statusKey))
-            .filter(Boolean) as ResultStatus[]
-        : [];
-    const selectedClasses: string[] = classParam
-        ? (classParam.split("~") as string[])
-        : [];
-
-    const handleStatusChange = (statuses: ResultStatus[]) => {
-        const params = new URLSearchParams(searchParams);
-
-        if (statuses.length > 0) {
-            const keys = statuses.map(
-                s => StatusService.get(s)?.key
-            ).filter(Boolean);
-
-            params.set("status", keys.join("~"));
-        } else {
-            params.delete("status");
-        }
-
-        setSearchParams(params);
-    };
-
-    const handleClassChange = (classes: string[]) => {
-        const params = new URLSearchParams(searchParams);
-
-        if (classes.length > 0) {
-            params.set("class", classes.join("~"));
-        } else {
-            params.delete("class");
-        }
-
-        setSearchParams(params);
-    };
-
-    const handleConfigurationMethodsChecked = (value: any) => {
-        setConfigurationMethodsChecked(value);
-    };
-
-    const handleDelete = (value: string | ResultStatus, type: "status" | "class") => {
-        setSearchParams(prev => {
-            const params = new URLSearchParams(prev);
-
-            if (type === "status") {
-                const updated = selectedStatuses.filter(s => s !== value);
-                if (updated.length > 0) {
-                    const keys = updated
-                        .map(s => StatusService.get(s)?.key)
-                        .filter(Boolean);
-                    params.set("status", keys.join("~"));
-                } else {
-                    params.delete("status");
-                }
-            }
-
-            if (type === "class") {
-                const updated = selectedClasses.filter(c => c !== value);
-                if (updated.length > 0) {
-                    params.set("class", updated.join("~"));
-                } else {
-                    params.delete("class");
-                }
-            }
-
-            return params;
-        });
-    };
-
+    const {
+        statusMenuItems,
+        classMenuItems,
+        selectedStatuses,
+        selectedClasses,
+        handleStatusChange,
+        handleClassChange,
+        configurationMethodsChecked,
+        handleConfigurationMethodsChecked,
+        handleDelete,
+        handleClearAllClick
+    } = useTestListFilters();
 
     return (
         <Box
@@ -113,7 +39,12 @@ const TestListPage = () => {
                 </Grid>
                 <Grid size={3}>
                     <MultiSelectInput label="Class" values={selectedClasses} onChange={handleClassChange}
-                                      menuItems={classMenuItems}/>
+                                      menuItems={classMenuItems}
+                                      renderValue={(selected: string[]) => {
+                                          if (!selected?.length) return "";
+                                          if (selected.length === 1) return "1 class selected";
+                                          return `${selected.length} classes selected`;
+                                      }}/>
                 </Grid>
                 <Grid size={5}>
                     <TextField
@@ -136,22 +67,10 @@ const TestListPage = () => {
                                       label="Show configuration methods"/>
                 </Grid>
                 <Grid size={12}>
-                    <Stack direction="row" spacing={1}>
-                        {selectedStatuses.map((status) => {
-                            const statusInformation = StatusService.get(status);
-                            if (!statusInformation) return null;
-
-                            return (
-                                <ReportChip key={status} label={statusInformation.label} type="status"
-                                    handleDelete={() => handleDelete(status, "status")}
-                                />
-                            );
-                        })}
-                        {selectedClasses.map((rClass) => (
-                            <ReportChip key={rClass} label={rClass} type="class"
-                                        handleDelete={() => handleDelete(rClass, "class")}/>
-                        ))}
-                    </Stack>
+                    <SelectedFiltersChips selectedStatuses={selectedStatuses}
+                                          selectedClasses={selectedClasses}
+                                          handleDelete={handleDelete}
+                                          handleClearAllClick={handleClearAllClick}/>
 
                 </Grid>
             </Grid>
