+++ report-ng/app-react/src/widgets/select-input/status-select-input.tsx
@@ -8,13 +8,15 @@ import type {ResultStatus} from "../../model/status-service";
 
 type SelectInputProps = {
     label: string
-    selectedStatuses?: ResultStatus[] | [],
+    selectedStatuses?: ResultStatus[],
     onChange: (value: ResultStatus[]) => void;
     menuItems: ResultStatus[];
     sx?: SxProps<Theme>;
 }
 
 const StatusSelectInput = ({label, selectedStatuses, onChange, menuItems, sx}: SelectInputProps) => {
+
+    const availableMenuItems = menuItems.filter(status => !selectedStatuses?.includes(status));
 
     return (
         <Box sx={sx}>
@@ -28,8 +30,19 @@ const StatusSelectInput = ({label, selectedStatuses, onChange, menuItems, sx}: S
                         onChange(e.target.value as ResultStatus[])
                     }
                     sx={{height: "56px"}}
+                    renderValue={(selected: ResultStatus[]) => {
+                        if (!selected?.length) return "";
+                        if (selected.length === 1) return "1 status selected";
+                        return `${selected.length} statuses selected`;
+                    }}
                 >
-                    {menuItems.map(status => {
+                    {availableMenuItems.length === 0 && (
+                        <MenuItem disabled>
+                            <em>All status selected</em>
+                        </MenuItem>
+                    )}
+                    {availableMenuItems
+                        .map(status => {
                         const statusInformation = StatusService.get(status);
                         if (!statusInformation) return null;
